requirements.txt
flask
python-dotenv
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-ollama
langchain-text-splitters
faiss-cpu
pypdf
python-docx
openai
torch
sentence-transformers